Struct llama::ExecutionEngine

pub struct ExecutionEngine<'a>(_, _, _);

An execution engine can be used to execute JIT compiled code

Implementations

impl<'a> ExecutionEngine<'a>

pub fn new(module: Module<'a>) -> Result<ExecutionEngine<'a>, Error>

Create a new execution engine using LLVMCreateExectionEngineForModule

pub fn new_jit(
module: Module<'a>,
opt: usize
) -> Result<ExecutionEngine<'a>, Error>

Create new JIT compiler with optimization level

pub unsafe fn function<T: 'a + Copy>(
&self,
name: impl AsRef<str>
) -> Result<T, Error>

Get a function from within the execution engine

Safety

This function does nothing to ensure that the function actually matches the type you give it

pub unsafe fn global_value<T>(
&self,
name: impl AsRef<str>
) -> Result<Value<'a>, Error>

Get a pointer to a global value from within the execution engine

Safety

This function does nothing to ensure that the function actually matches the type you give it

pub unsafe fn global<T: 'a>(
&self,
global: impl AsRef<Value<'a>>
) -> Result<&mut T, Error>

Get a pointer to a global from within the execution engine

Safety

This function does nothing to ensure that the function actually matches the type you give it

pub fn run_static_constructors(&self)

Run static constructors

pub fn run_static_destructors(&self)

Run static destructors

pub fn add_global_mapping<T: 'a>(
&mut self,
global: impl AsRef<Value<'a>>,
data: *mut T
)

Add mapping between global value and a local object

pub fn target_data(&self) -> Result<TargetData<'_>, Error>

Get target data

pub fn module(&self) -> &Module<'a>

Get handle to underlying module

pub fn module_mut(&mut self) -> &mut Module<'a>

Get mutable handle to underlying module

pub fn into_module(self) -> Result<Module<'a>, Error>

Get inner module