Struct llama::TargetData

pub struct TargetData<'a>(_, _);

LLVMTargetData wrapper

Implementations

impl<'a> TargetData<'a>

pub fn new(s: impl AsRef<str>) -> Result<TargetData<'a>, Error>

Create new target data with the given triple

pub fn string_rep(&self) -> Message

Return a string representation

pub fn byte_order(&self) -> ByteOrder

Get configured byte order

pub fn pointer_size(&self) -> usize

Get configured pointer size

pub fn pointer_size_for_address_space(&self, addr_space: usize) -> usize

Get pointer size for the given address space

pub fn int_ptr_type(&self, ctx: &Context<'a>) -> Result<Type<'a>, Error>

Get the type used for integer pointers

pub fn int_ptr_type_for_address_space(
&self,
ctx: &Context<'a>,
addr_space: usize
) -> Result<Type<'a>, Error>

Get the type used for integer pointers in the given address space

pub fn size_of_type_in_bits(&self, t: impl AsRef<Type<'a>>) -> usize

Get the size of a type in bits

pub fn store_size_of_type(&self, t: impl AsRef<Type<'a>>) -> usize

Get the storage size of a type

pub fn abi_size_of_type(&self, t: impl AsRef<Type<'a>>) -> usize

Get the ABI size of a type

pub fn abi_alignment_of_type(&self, t: impl AsRef<Type<'a>>) -> usize

Get the ABI alignment of a type

pub fn call_frame_alignment_of_type(&self, t: impl AsRef<Type<'a>>) -> usize

Get the call frame alignment of a type

pub fn preferred_alignment_of_type(&self, t: impl AsRef<Type<'a>>) -> usize

Get type’s preferred alignment

pub fn preferred_alignment_of_global(&self, t: impl AsRef<Value<'a>>) -> usize

Get preferred alignment for a global value

pub fn struct_element_at_offset(
&self,
t: impl AsRef<Type<'a>>,
offset: usize
) -> usize

Get the index of the struct element at the given offset

pub fn struct_offset_of_element(
&self,
t: impl AsRef<Type<'a>>,
index: usize
) -> usize

Get the offset of the element at the given index