Trait llama::PassManager

pub trait PassManager: LLVM<LLVMPassManager> {
    type Kind;
    fn run(&self, f: &Self::Kind) -> bool;

    fn add(&self, transforms: impl AsRef<[Transform]>) { ... }
}

PassManager trait is used to define common functionality between the two types of PassManagers

Associated Types

type Kind

Kind is used to designate the kind of value that can be optimized using this PassManager

Required methods

fn run(&self, f: &Self::Kind) -> bool

Run configured optimization passes

Provided methods

fn add(&self, transforms: impl AsRef<[Transform]>)

Add optimization passes