Trait llama::PassManager

pub trait PassManager: LLVM<LLVMPassManager> {
    type Kind;

    fn run(&self, f: &Self::Kind) -> bool;

    unsafe fn add(&self, transforms: impl AsRef<[Transform]>) { ... }
}

PassManager trait is used to define common functionality between the two types of PassManagers

Required Associated Types

type Kind

Kind is used to designate the kind of value that can be optimized using this PassManager

Required Methods

fn run(&self, f: &Self::Kind) -> bool

Run configured optimization passes

Provided Methods

unsafe fn add(&self, transforms: impl AsRef<[Transform]>)

Add optimization passes

Safety

This is marked as unsafe because Transform functions are unsafe